--- a/src/besseli.jl
+++ b/src/besseli.jl
@@ -11,18 +11,14 @@ http://www.advanpix.com/2015/11/12/rational-approximations-for-the-modified-bess
 Actual coefficients used are from the boost math library.
 https://github.com/boostorg/math/tree/develop/include/boost/math/special_functions/detail
 =#
-
 function besseli0(x::T) where T <: Union{Float32, Float64}
-    T == Float32 ? branch = 50 : branch = 500
     x = abs(x)
     if x < 7.75
         a = x * x / 4
-        return muladd(a, evalpoly(a, P1_i0(T)), 1)
-    elseif x < branch
-        return exp(x) * evalpoly(inv(x), P2_i0(T)) / sqrt(x)
+        return muladd(a, evalpoly(a, besseli0_small_coefs(T)), 1)
     else
         a = exp(x / 2)
-        s = a * evalpoly(inv(x), P3_i0(T)) / sqrt(x)
+        s = a * evalpoly(inv(x), besseli0_med_coefs(T)) / sqrt(x)
         return a * s
     end
 end
@@ -31,11 +27,11 @@ function besseli0x(x::T) where T <: Union{Float32, Float64}
     x = abs(x)
     if x < 7.75
         a = x * x / 4
-        return muladd(a, evalpoly(a, P1_i0(T)), 1) * exp(-x)
+        return muladd(a, evalpoly(a, besseli0_small_coefs(T)), 1) * exp(-x)
     elseif x < branch
-        return evalpoly(inv(x), P2_i0(T)) / sqrt(x)
+        return evalpoly(inv(x), besseli0_med_coefs(T)) / sqrt(x)
     else
-        return evalpoly(inv(x), P3_i0(T)) / sqrt(x)
+        return evalpoly(inv(x), besseli0_large_coefs(T)) / sqrt(x)
     end
 end
 function besseli1(x::Float32)
@@ -43,11 +39,11 @@ function besseli1(x::Float32)
     z = abs(x)
     if z < 7.75
         a = z * z / 4
-        inner = (one(T), T(0.5), evalpoly(a, P1_i1(T)))
+        inner = (one(T), T(0.5), evalpoly(a, besseli1_small_coefs(T)))
         z = z * evalpoly(a, inner) / 2
     else
         a = exp(z / 2)
-        s = a * evalpoly(inv(z), P2_i1(T)) / sqrt(z)
+        s = a * evalpoly(inv(z), besseli1_med_coefs(T)) / sqrt(z)
         z =  a * s
     end
     if x < zero(x)
@@ -60,13 +56,11 @@ function besseli1(x::Float64)
     z = abs(x)
     if z < 7.75
         a = z * z / 4
-        inner = (one(T), T(0.5), evalpoly(a, P1_i1(T)))
+        inner = (one(T), T(0.5), evalpoly(a, besseli1_small_coefs(T)))
         z = z * evalpoly(a, inner) / 2
-    elseif z < 500
-        return exp(z) * evalpoly(inv(z), P2_i1(T)) / sqrt(z)
     else
         a = exp(z / 2)
-        s = a * evalpoly(inv(z), P3_i1(T)) / sqrt(z)
+        s = a * evalpoly(inv(z), besseli1_med_coefs(T)) / sqrt(z)
         z =  a * s
     end
     if x < zero(x)
@@ -79,10 +73,10 @@ function besseli1x(x::Float32)
     z = abs(x)
     if z < 7.75
         a = z * z / 4
-        inner = (one(T), T(0.5), evalpoly(a, P1_i1(T)))
+        inner = (one(T), T(0.5), evalpoly(a, besseli1_small_coefs(T)))
         z = z * evalpoly(a, inner) / 2 * exp(-z)
     else
-        z =  evalpoly(inv(z), P2_i1(T)) / sqrt(z)
+        z =  evalpoly(inv(z), besseli1_med_coefs(T)) / sqrt(z)
     end
     if x < zero(x)
         z = -z
@@ -94,12 +88,12 @@ function besseli1x(x::Float64)
     z = abs(x)
     if z < 7.75
         a = z * z / 4
-        inner = (one(T), T(0.5), evalpoly(a, P1_i1(T)))
+        inner = (one(T), T(0.5), evalpoly(a, besseli1_small_coefs(T)))
         z = z * evalpoly(a, inner) / 2 * exp(-z)
     elseif z < 500
-        z = evalpoly(inv(z), P2_i1(T)) / sqrt(z)
+        z = evalpoly(inv(z), besseli1_med_coefs(T)) / sqrt(z)
     else
-        z = evalpoly(inv(z), P3_i1(T)) / sqrt(z)
+        z = evalpoly(inv(z), besseli1_large_coefs(T)) / sqrt(z)
     end
     if x < zero(x)
         z = -z