--- a/src/besseli.jl
+++ b/src/besseli.jl
@@ -1,16 +1,49 @@
-# Modified Bessel functions of the first kind of order zero and one
-# besseli0, besseli1
-# Scaled modified Bessel functions of the first kind of order zero and one
-# besseli0x, besselix
+#    Modified Bessel functions of the first kind of order zero and one
+#                       besseli0, besseli1
+#
+#    Scaled modified Bessel functions of the first kind of order zero and one
+#                       besseli0x, besselix
+#
+#    Calculation of besseli0 is done in two branches using polynomial approximations [1]
+#
+#    Branch 1: x < 7.75 
+#              besseli0 = [x/2]^2 P16([x/2]^2)
+#    Branch 2: x >=
+#              sqrt(x) * exp(-x) * besseli0(x) = P22(1/x)
+#    where P16 and P22 are a 16 and 22 degree polynomial respectively.
+#
+#    Remez.jl is then used to approximate the polynomial coefficients of
+#    P22(y) = sqrt(1/y) * exp(-inv(y)) * besseli0(inv(y))
+#    N,D,E,X = ratfn_minimax(g, (1/1e6, 1/7.75), 21, 0)
+#
+#    A third branch is used for scaled functions for large values
+#
+#
+#    Calculation of besseli1 is done in two branches using polynomial approximations [2]
+#
+#    Branch 1: x < 7.75 
+#              besseli1 = x / 2 * (1 + 1/2 * (x/2)^2 + (x/2)^4 * P13([x/2]^2)
+#    Branch 2: x >=
+#              sqrt(x) * exp(-x) * besseli0(x) = P22(1/x)
+#    where P13 and P22 are a 16 and 22 degree polynomial respectively.
+#
+#    Remez.jl is then used to approximate the polynomial coefficients of
+#    P13(y) = (besseli1(2 * sqrt(y)) / sqrt(y) - 1 - y/2) / y^2
+#    N,D,E,X = ratfn_minimax(g, (1/1e6, 1/7.75), 21, 0)
+#
+#    A third branch is used for scaled functions for large values
+#
+#    Horner's scheme is then used to evaluate all polynomials.
+# 
+# [1] "Rational Approximations for the Modified Bessel Function of the First Kind 
+#     - I0(x) for Computations with Double Precision" by Pavel Holoborodko     
+# [2] "Rational Approximations for the Modified Bessel Function of the First Kind 
+#     - I1(x) for Computations with Double Precision" by Pavel Holoborodko        
+"""
+    besseli0(x::T) where T <: Union{Float32, Float64}
 
-#=
-Approximate forms used here are given in
-"Rational Approximations for the Modified Bessel Function of the First Kind - I1(x) for Computations with Double Precision"
-by Pavel Holoborodko, 
-http://www.advanpix.com/2015/11/12/rational-approximations-for-the-modified-bessel-function-of-the-first-kind-i1-for-computations-with-double-precision/
-Actual coefficients used are from the boost math library.
-https://github.com/boostorg/math/tree/develop/include/boost/math/special_functions/detail
-=#
+Modified Bessel function of the first kind of order zero, ``I_0(x)``.
+"""
 function besseli0(x::T) where T <: Union{Float32, Float64}
     x = abs(x)
     if x < 7.75
@@ -22,6 +55,11 @@ function besseli0(x::T) where T <: Union{Float32, Float64}
         return a * s
     end
 end
+"""
+    besseli0x(x::T) where T <: Union{Float32, Float64}
+
+Scaled modified Bessel function of the first kind of order zero, ``I_0(x)*e^{-x}``.
+"""
 function besseli0x(x::T) where T <: Union{Float32, Float64}
     T == Float32 ? branch = 50 : branch = 500
     x = abs(x)
@@ -34,6 +72,11 @@ function besseli0x(x::T) where T <: Union{Float32, Float64}
         return evalpoly(inv(x), besseli0_large_coefs(T)) / sqrt(x)
     end
 end
+"""
+    besseli1(x::T) where T <: Union{Float32, Float64}
+
+Modified Bessel function of the first kind of order one, ``I_1(x)``.
+"""
 function besseli1(x::T) where T <: Union{Float32, Float64}
     z = abs(x)
     if z < 7.75
@@ -50,6 +93,11 @@ function besseli1(x::T) where T <: Union{Float32, Float64}
     end
     return z
 end
+"""
+    besseli1x(x::T) where T <: Union{Float32, Float64}
+
+Scaled modified Bessel function of the first kind of order one, ``I_1(x)*e^{-x}``.
+"""
 function besseli1x(x::T) where T <: Union{Float32, Float64}
     T == Float32 ? branch = 50 : branch = 500
     z = abs(x)