--- a/test/besseli_test.jl
+++ b/test/besseli_test.jl
@@ -1,5 +1,5 @@
 # general array for testing input to SpecialFunctions.jl
-x32 = 1e-6:0.01:50.0
+x32 = 1e-6:0.01:60.0
 x64 = 1e-6:0.01:600.0
 ### Tests for besseli0
 i032_SpecialFunctions = SpecialFunctions.besseli.(0, x32)