@heltonmc 2022-01-14
So tested against `ArbNumerics.jl` using 50 digits of precision for 10000 evenly spaced points between 0 and 100 the mean and max relative errors are as follows for `besseli0(x)`:

Old versions
Mean relative error: 1.00800350195715708e-16
Max relative error: 3.75163868123101539944e-16

New version
Mean relative error: 1.184275524171108551677614992624e-16
Max relative error: 4.036293772497829275701727546031e-16

So a very slight increase in error.
Is there a built in function to calculate ULP from the residuals?