@heltonmc 2022-01-15
I've recalculated all the coefficients used in these routines. These new weights signficinatly improve errors at large input arguments. I've also added new documentation for internal and external api
 
## `besseli0`
![Screenshot from 2022-01-15 15-21-50](https://user-images.githubusercontent.com/57471570/149640945-2d5844c9-30f0-4890-977f-c8ea1a8d3ff5.png)


## `besseli1`
![Screenshot from 2022-01-15 16-25-37](https://user-images.githubusercontent.com/57471570/149640949-5b7e1f7d-cefa-48b2-9d77-350bc8dcc0bc.png)

I think this is ready to merge now.