@oscardssmith 2022-01-15
One question: Why did you only use 75 digits for Arb? I would recommend using at least 256 since it's basically free.