@heltonmc 2022-01-16
I wasn't very clear and for repeatability purposes let me post an example code.
```julia
 bessi1(x) = besseli(1, ArbFloat(x, digits=75))
g(y) = inv(y) * (bessi(2*sqrt(y)) - 1)
N,D,E,X = ratfn_minimax(g, (1e-18, (7.75/2)^2), 6, 0)
```
I just went and reran with digits=256 and they match with digits=75 to about 49 digits.
So for example Ill get something like
```julia
# digits = 75
julia> N
7-element Vector{BigFloat}:
 0.08333336088599108314227728172065098704108098333505801124750081760507945728818461
.
.
# digits = 256
julia> N
7-element Vector{BigFloat}:
 0.0833333608859910831422772817206509870410809833845858298944772410690681516396821
.
.
```

I then will convert them to Float64 or Float32. I think this is ok for Float64. One thing is do we want to also give Float128 or DoubleFloats or higher precision implementations? I do think this is important.... Should we just take on the DoubleFloats or Quadmath dependency and define constants for this? This shouldn't be too hard.

It's also important to document this issue https://github.com/JeffreySarnoff/ArbNumerics.jl/issues/48 with `ArbNumerics.jl`. This makes the evaluation quite slow but should always be careful that your true functions are calculated correctly..